# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "openrouter · servicio",
    "version": "0.8.0",
    "summary": "Pasarela propia sobre OpenRouter y el router de Hugging Face, pensada para ser llamada por otras aplicaciones.",
    "description": "Esto no es una aplicación con interfaz: es un conjunto de servicios HTTP. La clave de OpenRouter vive solo en el backend y nunca sale de él. Cada aplicación consumidora tiene su propia clave, que se da de alta y se revoca sin tocar el despliegue.\n\nEl contrato sigue el de la API de OpenAI donde existe equivalente, así que cualquier cliente de OpenAI funciona cambiando la URL base. Lo propio del servicio va en rutas propias y en cabeceras `X-...`. Una ruta publicada no se rompe: lo incompatible iría a `/v2/`.\n\nLa versión de este documento acompaña a la etapa entregada: `0.6.0` es la etapa 6; `0.6.2` amplía `/v1/models` con descripción, precios completos, modalidades de salida y parámetros, sin cambiar nada de lo publicado; `0.6.4` entrega la etapa 9 (catálogo completo de OpenRouter y compra en el router de Hugging Face con el prefijo `hf:`), también sin cambiar nada: solo añade campos, parámetros y códigos. Las etapas 7 y 8 siguen pendientes.",
    "license": {
      "name": "Uso interno"
    }
  },
  "servers": [
    {
      "url": "https://apisor.oracle402.com",
      "description": "Producción (VPS)"
    },
    {
      "url": "https://openrouter-npiobject-labs.fly.dev",
      "description": "Previsualización de main (Fly.io)"
...
  "security": [
    {
      "claveDeServicio": []
    }
  ],
...
  "paths": {
...
    "/v1/estado": {
      "get": {
        "tags": [
          "Servicio"
        ],
        "summary": "Estado del servicio y de la clave",
        "description": "Comprueba de verdad que el backend habla con OpenRouter, y devuelve la etiqueta de la clave, su uso y su límite tal como los da el proveedor. No gasta crédito, así que sirve para vigilancia periódica.",
        "responses": {
          "200": {
            "description": "Conexión confirmada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estado"
                },
                "example": {
                  "ok": true,
                  "build": "87e6a89",
                  "modelo_defecto": "openai/gpt-oss-120b",
                  "modelo_multimodal": "google/gemini-3.1-flash-lite",
                  "clave": {
                    "label": "openrouter",
                    "usage": 0.0123,
                    "limit": 5,
                    "is_free_tier": false
                  },
                  "almacen": "sqlite",
                  "registros": 128,
                  "identidad": {
                    "nombre": "presupuestos",
                    "app_id": "app_27b1d11e45",
                    "administracion": false
                  },
                  "upstreams": {
                    "openrouter": true,
                    "hf": true
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "503": {
            "$ref": "#/components/responses/SinConfigurar"
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "Modelos"
        ],
        "summary": "Catálogo de modelos",
        "description": "El catálogo del upstream elegido con el formato de la API de OpenAI, más los campos que hacen falta para elegir: precio en dólares por millón de tokens, ventana de contexto, modalidades de entrada y de salida, si admite herramientas o salida estructurada y, desde 0.6.2, la descripción del proveedor, los precios completos (imagen, petición, caché, audio) y la lista de parámetros.\n\nDesde 0.6.4 el catálogo de OpenRouter es el completo (`output_modalities=all`): además de los modelos que escriben texto trae los de imagen, vídeo, voz, transcripción y embeddings, que antes quedaban ocultos. Cada modelo dice su `upstream` y conserva `canonical_slug` y `hugging_face_id`. Con `salida=<modalidad>` se pide solo lo que interesa; sin él, todo.\n\nCon `upstream=hf` se devuelve el catálogo del router de Hugging Face normalizado a la misma forma: una entrada `hf:<id>` por modelo, con el precio del host más barato, y una `hf:<id>:<host>` por cada host que lo sirve con su precio, contexto, latencia hasta el primer token y tokens por segundo. Todos los ids se mandan tal cual en `model`. Sin `upstream`, el de OpenRouter, como siempre.\n\nCada catálogo se cachea una hora en memoria. Si el upstream no responde y hay copia caducada, se sirve la caducada antes que un error: para elegir modelo, una lista de ayer es mejor que nada. La cabecera `X-Cache` dice de dónde salió.",
        "parameters": [
          {
            "name": "texto",
            "in": "query",
            "description": "Filtra por texto en el identificador o el nombre.",
            "schema": {
              "type": "string"
            },
            "example": "gemini"
          },
          {
            "name": "proveedor",
            "in": "query",
            "description": "Filtra por proveedor, que es el tramo anterior a la barra del identificador.",
            "schema": {
              "type": "string"
            },
            "example": "anthropic"
          },
          {
            "name": "contexto_min",
            "in": "query",
            "description": "Descarta los modelos con menos contexto que este número de tokens.",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "example": 200000
          },
          {
            "name": "gratis",
            "in": "query",
            "description": "Con `1`, `true` o `si`, solo modelos de precio cero. Cualquier otro valor no filtra.",
            "schema": {
              "type": "string"
            },
            "example": "1"
          },
          {
            "name": "refrescar",
            "in": "query",
            "description": "Con `1`, `true` o `si`, ignora la caché y vuelve a pedir el catálogo.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salida",
            "in": "query",
            "description": "Solo los modelos que producen esa modalidad: `text`, `image`, `video`, `speech`, `transcription` o `embeddings`. Sin él, todas.",
            "schema": {
              "type": "string",
              "enum": [
                "text",
                "image",
                "video",
                "speech",
                "transcription",
                "embeddings"
              ]
            },
            "example": "image"
          },
          {
            "name": "upstream",
            "in": "query",
            "description": "De quién es el catálogo: `openrouter` (por defecto), `hf` (router de Hugging Face) o `rq` (Requesty). Sin la clave de ese upstream en el despliegue, `503 upstream_sin_configurar`. Un valor desconocido es `400 upstream_desconocido`.",
            "schema": {
              "type": "string",
              "enum": [
                "openrouter",
                "hf",
                "rq"
              ],
              "default": "openrouter"
            },
            "example": "hf"
          },
          {
            "name": "region",
            "in": "query",
            "description": "Solo los modelos que publican que procesan en esa región (`eu`, `us`, `global`, `uk`, `sg`, `ap`). Hoy solo Requesty lo publica: en los demás upstreams no queda ninguno.",
            "schema": {
              "type": "string"
            },
            "example": "eu"
          },
          {
            "name": "sin_retencion",
            "in": "query",
            "description": "`1`: solo los que publican que ni guardan las peticiones ni entrenan con ellas (`retencion: false` y `entrena: false`). Hoy solo Requesty lo publica.",
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de modelos. `total` es el catálogo entero, antes de aplicar filtros.",
            "headers": {
              "X-Cache": {
                "description": "`miss` recién traído de OpenRouter, `hit` de la caché vigente, `stale` copia caducada porque OpenRouter no respondió.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "miss",
                    "hit",
                    "stale"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaDeModelos"
                },
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "id": "google/gemini-2.5-flash-lite",
                      "object": "model",
                      "created": 1752000000,
                      "owned_by": "google",
                      "nombre": "Google: Gemini 2.5 Flash Lite",
                      "contexto": 1048576,
                      "entrada": 0.1,
                      "salida": 0.4,
                      "gratis": false,
                      "herramientas": true,
                      "json": true,
                      "modalidades": [
                        "text",
                        "image"
                      ],
                      "descripcion": "Gemini 2.5 Flash Lite es la variante más rápida y barata de la familia Flash.",
                      "precios": {
                        "imagen": 0,
                        "peticion": 0,
                        "cache_lectura": 0.025,
                        "cache_escritura": 0,
                        "audio": 0,
                        "razonamiento": 0,
                        "busqueda_web": 0,
                        "imagen_salida": 0
                      },
                      "modalidades_salida": [
                        "text"
                      ],
                      "parametros": [
                        "tools",
                        "structured_outputs",
                        "temperature",
                        "max_tokens"
                      ],
                      "max_salida": 65535,
                      "moderado": false,
                      "upstream": "openrouter",
                      "canonical_slug": "google/gemini-2.5-flash-lite-20250617",
                      "hugging_face_id": null,
                      "primer_token_ms": null,
                      "tokens_por_segundo": null
                    },
                    {
                      "id": "hf:openai/gpt-oss-120b:groq",
                      "object": "model",
                      "created": 1754346786,
                      "owned_by": "openai",
                      "nombre": "openai/gpt-oss-120b · groq",
                      "contexto": 131072,
                      "entrada": 0.15,
                      "salida": 0.75,
                      "gratis": false,
                      "herramientas": true,
                      "json": true,
                      "modalidades": [
                        "text"
                      ],
                      "descripcion": "",
                      "precios": {
                        "imagen": 0,
                        "peticion": 0,
                        "cache_lectura": 0,
                        "cache_escritura": 0,
                        "audio": 0,
                        "razonamiento": 0,
                        "busqueda_web": 0,
                        "imagen_salida": 0
                      },
                      "modalidades_salida": [
                        "text"
                      ],
                      "parametros": [],
                      "max_salida": null,
                      "moderado": false,
                      "upstream": "hf",
                      "canonical_slug": null,
                      "hugging_face_id": "openai/gpt-oss-120b",
                      "primer_token_ms": 228,
                      "tokens_por_segundo": 416.2
                    }
                  ],
                  "total": 649,
                  "antiguedad_minutos": 12,
                  "upstream": "openrouter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PeticionInvalida"
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "502": {
            "$ref": "#/components/responses/FalloDeProveedor"
          },
          "503": {
            "$ref": "#/components/responses/SinConfigurar"
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Consulta a un modelo",
        "description": "Proxy fino sobre OpenRouter, y desde 0.6.4 también sobre el router de Hugging Face, con el contrato de OpenAI. A quién va lo decide el prefijo de `model`: `hf:<id>` sale hacia Hugging Face sin el prefijo (`hf:meta-llama/Llama-3.3-70B-Instruct:groq` pide `meta-llama/Llama-3.3-70B-Instruct:groq`, donde `:groq` es el host; `:cheapest` y `:fastest` dejan que elija el router); todo lo demás va a OpenRouter exactamente como siempre. Sin `HF_TOKEN` en el despliegue, `hf:` responde `503 upstream_sin_configurar`.\n\nEl cuerpo se reenvía tal cual salvo tres retoques: el prefijo, si no traes `model` se rellena con el modelo por defecto del servicio (`modelo_defecto` de `/v1/estado`), o con el multimodal (`modelo_multimodal`) si algún mensaje lleva una parte que no es texto; con el de texto, y si no mandas `reasoning`, se añade `{\"effort\":\"low\"}` para que el razonamiento no se coma el `max_tokens`; y `stream: true` se rechaza con `400` porque el streaming llega en la etapa 7.\n\nLa respuesta es la del upstream sin tocar, incluido el bloque `usage`, que es lo que permite saber qué costó la llamada. En Hugging Face el coste queda `estimado` con los precios de su catálogo (los del host pedido, o los del más barato si no se pidió ninguno) y no se reconcilia: el router no tiene `/generation`.\n\nToda llamada que llega a salir queda anotada en el registro de uso, vaya bien o mal, y su identificador vuelve en la cabecera `X-Uso-Id`. Con él se consulta después lo que costó, en `/v1/uso/{id}`.\n\nVaya a un upstream o a otro, con clave de aplicación la llamada pasa antes por los mismos tres controles: el cortacircuitos de bucles, la cuota por minuto y el presupuesto. La clave de administración no tiene ninguno.\n\nDesde 0.7.0 `model` admite también `rq:<id de Requesty>` (`rq:vertex/gemini-3.1-flash-lite@eu`), con el coste real que da Requesty en `usage.cost`. Y si OpenRouter no está (sin conexión, tiempo agotado, 5xx, 429 o 402), la consulta se repite automáticamente con el mismo modelo en Hugging Face y, si no, en Requesty (la variante de la UE que no retiene datos, la más barata). La respuesta dice quién contestó en `X-Upstream` y que hubo respaldo en `X-Respaldo`. Cada intento queda en el registro de uso. Para que una consulta nunca salga de OpenRouter, manda `X-Respaldo: no`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consulta"
              },
              "example": {
                "model": "google/gemini-2.5-flash-lite",
                "messages": [
                  {
                    "role": "user",
                    "content": "Resume en una línea qué es OpenRouter."
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta del modelo, con el formato de OpenAI.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta"
                },
                "example": {
                  "id": "gen-1757000000-abc",
                  "model": "google/gemini-2.5-flash-lite",
                  "object": "chat.completion",
                  "choices": [
                    {
                      "index": 0,
                      "finish_reason": "stop",
                      "message": {
                        "role": "assistant",
                        "content": "Una pasarela que da acceso a modelos de muchos proveedores con una sola clave."
                      }
                    }
                  ],
                  "usage": {
                    "prompt_tokens": 14,
                    "completion_tokens": 21,
                    "total_tokens": 35
                  }
                }
              }
            },
            "headers": {
              "X-Uso-Id": {
                "description": "Identificador del registro de uso de esta llamada. También viaja en las respuestas de error, cuando el fallo ocurrió tras salir hacia OpenRouter.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Presupuesto": {
                "description": "Aparece cuando el gasto del periodo ha pasado el umbral de aviso. No corta la llamada: es para que la aplicación sepa que se acerca al tope.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Upstream": {
                "description": "Quién respondió: `openrouter`, `hf` o `rq`.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "openrouter",
                    "hf",
                    "rq"
                  ]
                }
              },
              "X-Respaldo": {
                "description": "Solo si OpenRouter falló (sin conexión, tiempo agotado, 5xx, 429 o 402) y la consulta se repitió con el mismo modelo en otro upstream: el upstream que falló. En la petición, `X-Respaldo: no` apaga el respaldo para esa consulta.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PeticionInvalida"
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "402": {
            "description": "Presupuesto agotado, o crédito insuficiente en OpenRouter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "ok": false,
                  "error": {
                    "message": "Esta aplicación lleva 5.10 $ gastados y su presupuesto por mes es de 5.00 $.",
                    "code": "presupuesto_agotado",
                    "type": "insufficient_quota"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NoEncontrado"
          },
          "429": {
            "description": "Cuota por minuto superada, o la misma consulta repetida demasiadas veces seguidas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "ok": false,
                  "error": {
                    "message": "La misma consulta se ha repetido 5 veces en 60 segundos. Parece un bucle, así que se corta para no gastar crédito.",
                    "code": "bucle",
                    "type": "rate_limit_error"
                  }
                }
              }
            }
          },
          "502": {
            "$ref": "#/components/responses/FalloDeProveedor"
          },
          "503": {
            "$ref": "#/components/responses/SinConfigurar"
          }
        },
        "parameters": [
          {
            "name": "X-Operacion",
            "in": "header",
            "required": false,
            "description": "Identificador del trabajo de negocio al que pertenece la llamada, por ejemplo el número de presupuesto. Se guarda en el registro para agrupar después varias llamadas del mismo trabajo. Hasta 120 caracteres.",
            "schema": {
              "type": "string"
            },
            "example": "presupuesto-2026-041"
          }
        ]
      }
    },
    "/v1/uso": {
      "get": {
        "tags": [
          "Uso"
        ],
        "summary": "Últimas llamadas",
        "description": "El registro de las últimas llamadas, de la más reciente a la más antigua.\n\nDesde la etapa 4 salen del histórico en disco, así que sobreviven a un reinicio. Para ventanas de tiempo y totales, `/v1/uso/resumen`; para llevarse los datos, `/v1/uso/exportar`.\n\nCon clave de aplicación solo salen sus propias llamadas; `total` sigue siendo el del servicio.",
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "description": "Cuántas devolver. Entre 1 y 1000; por defecto 50.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 50
            },
            "example": 10
          },
          {
            "name": "app",
            "in": "query",
            "description": "Con clave de administración, limita a esa aplicación. Con clave de aplicación se ignora: cada una ve solo lo suyo.",
            "schema": {
              "type": "string"
            },
            "example": "app_27b1d11e45"
          },
          {
            "name": "operacion",
            "in": "query",
            "description": "Solo las llamadas que llevaron ese valor en la cabecera `X-Operacion`: el trabajo de negocio (un presupuesto, un análisis) al que pertenecen.",
            "schema": {
              "type": "string"
            },
            "example": "presupuesto-2026-0417"
          },
          {
            "name": "upstream",
            "in": "query",
            "description": "Solo lo comprado en ese upstream: `openrouter` o `hf`. Un valor desconocido es `400 upstream_desconocido`.",
            "schema": {
              "type": "string",
              "enum": [
                "openrouter",
                "hf"
              ]
            },
            "example": "hf"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de registros. `total` es lo que hay guardado ahora mismo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "object": {
                      "type": "string",
                      "const": "list"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Registro"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                },
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "id": "u-1757673600123-7",
                      "fecha": "2026-09-12T10:40:00Z",
                      "modelo_pedido": "google/gemini-2.5-flash-lite",
                      "modelo_servido": "google/gemini-2.5-flash-lite",
                      "proveedor": "Google",
                      "tokens_entrada": 1000,
                      "tokens_salida": 500,
                      "tokens_razonamiento": 0,
                      "tokens_cache": 0,
                      "coste": 0.00042,
                      "coste_origen": "openrouter",
                      "latencia_ms": 1840,
                      "motivo_fin": "stop",
                      "estado": 200,
                      "id_openrouter": "gen-abc"
                    }
                  ],
                  "total": 1
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PeticionInvalida"
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          }
        }
      }
    },
    "/v1/uso/{id}": {
      "get": {
        "tags": [
          "Uso"
        ],
        "summary": "Una llamada concreta",
        "description": "El registro que corresponde a un `X-Uso-Id`.\n\nDesde la etapa 4 el histórico está en disco y no caduca.\n\nEl coste llega primero estimado con los precios del catálogo y, un par de segundos después, reconciliado con lo que factura OpenRouter. El campo `coste_origen` dice cuál de los dos estás viendo, así que una aplicación que quiera el dato bueno puede volver a preguntar.\n\nUna aplicación solo puede consultar registros suyos: los de otra responden como si no existieran.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "El identificador devuelto en `X-Uso-Id`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El registro.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registro"
                },
                "example": {
                  "id": "u-1757673600123-7",
                  "fecha": "2026-09-12T10:40:00Z",
                  "modelo_pedido": "google/gemini-2.5-flash-lite",
                  "modelo_servido": "google/gemini-2.5-flash-lite",
                  "proveedor": "Google",
                  "tokens_entrada": 1000,
                  "tokens_salida": 500,
                  "tokens_razonamiento": 0,
                  "tokens_cache": 0,
                  "coste": 0.00042,
                  "coste_origen": "openrouter",
                  "latencia_ms": 1840,
                  "motivo_fin": "stop",
                  "estado": 200,
                  "id_openrouter": "gen-abc"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NoAutorizado"
          },
          "404": {
            "description": "No hay registro con ese id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "ok": false,
                  "error": {
                    "message": "No hay registro de uso con id u-1-2.",
                    "code": "uso_desconocido",
                    "type": "not_found_error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "claveDeServicio": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clave en `Authorization: Bearer <clave>`.\n\nHay dos tipos. La **clave de administración** es la del despliegue y puede con todo, incluidas las rutas de `/v1/apps`. Las **claves de aplicación** se dan de alta en `POST /v1/apps`, se revocan una a una y solo ven su propio gasto.\n\nNunca la pongas en código de navegador: llama desde el backend de tu aplicación. Las cuotas y presupuestos por aplicación llegan en la etapa 6."
      }
    },
...
    "responses": {
      "NoAutorizado": {
        "description": "Falta la clave de servicio o no es válida.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "ok": false,
              "error": {
                "message": "Falta la cabecera Authorization: Bearer <clave de servicio>.",
                "code": "sin_clave",
                "type": "authentication_error"
              }
            }
          }
        }
      },
...
    }
  },
...
}